Lua Files & Script Manual
In this file could be found is the parametric model of the Internal Gear Pump scripted in Lua. To this provided is the instruction Manual which summarizes the model parts and also the added instruction for generating the .STL files of every component and also the G Code with the slicing instructions required for 3D printing.